chore(test) create test/e2e infrastructure and move the code there
